refactor(optim)!: Split GenericLocalSearchOptimizer into loop + wrapper
Rename the trial/accept loop to LocalSearchLoop (now under optim::search_loop) and add a new GenericLocalSearchOptimizer that owns a TransitionHandler and implements LocalSearchOptimizer.

All annealing optimizers keep working unchanged through the rename. The new wrapper exists for callers who want to drive a LocalSearchOptimizer through the standard trait interface with a stored handler.

BREAKING CHANGE: GenericLocalSearchOptimizer is no longer the raw loop. The raw loop is now LocalSearchLoop (3-arg constructor, optimize_with_handler). The new GenericLocalSearchOptimizer takes a 4th handler argument and implements LocalSearchOptimizer. StepResult moved from optim::generic to optim::search_loop (re-exported as crate::optim::StepResult).
