+++ src/optim/adaptive_annealing.rs
@@ -2,7 +2,7 @@ use std::num::NonZero;
 
 use ordered_float::NotNan;
 
-use super::{AdaptiveAnnealing, AdaptiveScheduler, GenericLocalSearchOptimizer, LocalSearchOptimizer,
+use super::{AdaptiveAnnealing, AdaptiveScheduler, LocalSearchLoop, LocalSearchOptimizer,
     tune_temperature,
 };
 use crate::{Duration, OptModel, callback::OptCallbackFn};
@@ -86,7 +86,7 @@ impl<M: OptModel<ScoreType = NotNan<f64>>> LocalSearchOptimizer<M> for AdaptiveA
     ) -> (M::SolutionType, M::ScoreType) {
         let handler =
             AdaptiveAnnealing::new(self.initial_beta, self.scheduler, self.update_frequency);
-        let opt = GenericLocalSearchOptimizer::new(self.patience, self.n_trials, self.return_iter);
+        let opt = LocalSearchLoop::new(self.patience, self.n_trials, self.return_iter);
         let (solution, score, _) = opt.optimize_with_handler(
             model,
             initial_solution,
