+++ src/optim/tsallis.rs
@@ -3,7 +3,7 @@ use std::num::NonZero;
 use ordered_float::NotNan;
 
 use super::{
-    AdaptiveScheduler, GenericLocalSearchOptimizer, LocalSearchOptimizer, TargetAccScheduleMode,
+    AdaptiveScheduler, LocalSearchLoop, LocalSearchOptimizer, TargetAccScheduleMode,
     TsallisAnnealing,
 };
 use crate::{Duration, OptModel, callback::OptCallbackFn};
@@ -97,7 +97,7 @@ impl<M: OptModel<ScoreType = NotNan<f64>>> LocalSearchOptimizer<M>
             self.scheduler,
             self.update_frequency,
         );
-        let opt = GenericLocalSearchOptimizer::new(self.patience, self.n_trials, self.return_iter);
+        let opt = LocalSearchLoop::new(self.patience, self.n_trials, self.return_iter);
         let (solution, score, _) = opt.optimize_with_handler(
             model,
             initial_solution,
